Move top toolbar to own file as ContextualToolbar
